Estimates and inference for sensitivity analyses

Computes point estimates, standard errors, and confidence interval bounds for (1) prop, the proportion of studies with true effect sizes above q (or below q for an apparently preventive yr) as a function of the bias parameters; (2) the minimum bias factor on the relative risk scale (Tmin) required to reduce to less than r the proportion of studies with true effect sizes more extreme than q; and (3) the counterpart to (2) in which bias is parameterized as the minimum relative risk for both confounding associations (Gmin).

confounded_meta(
  q,
  r = NA,
  muB = NA,
  sigB = 0,
  yr,
  vyr = NA,
  t2,
  vt2 = NA,
  CI.level = 0.95,
  tail = NA
)

Arguments

q
True effect size that is the threshold for "scientific significance"

r
For Tmin and Gmin, value to which the proportion of large effect sizes is to be reduced

muB
Mean bias factor on the log scale across studies

sigB
Standard deviation of log bias factor across studies

yr
Pooled point estimate (on log scale) from confounded meta-analysis

vyr
Estimated variance of pooled point estimate from confounded meta-analysis

t2
Estimated heterogeneity (tau^2) from confounded meta-analysis

vt2
Estimated variance of tau^2 from confounded meta-analysis

CI.level
Confidence level as a proportion

tail
above for the proportion of effects above q; below for the proportion of effects below q. By default, is set to above for relative risks above 1 and to below for relative risks below 1.

Details

To compute all three point estimates (prop, Tmin, and Gmin) and inference, all arguments must be non-NA. To compute only a point estimate for prop, arguments r, vyr, and vt2 can be left NA. To compute only point estimates for Tmin and Gmin, arguments muB, vyr, and vt2 can be left NA. To compute inference for all point estimates, vyr and vt2 must be supplied.

Examples

d = metafor::escalc(measure="RR", ai=tpos, bi=tneg,
ci=cpos, di=cneg, data=metafor::dat.bcg)

m = metafor::rma.uni(yi= d$yi, vi=d$vi, knha=FALSE,
                     measure="RR", method="DL" )
yr = as.numeric(m$b)  # metafor returns on log scale
vyr = as.numeric(m$vb)
t2 = m$tau2
vt2 = m$se.tau2^2

# obtaining all three estimators and inference
confounded_meta( q=log(0.90), r=0.20, muB=log(1.5), sigB=0.1,
                 yr=yr, vyr=vyr, t2=t2, vt2=vt2,
                 CI.level=0.95 )

#>   Value       Est        SE     CI.lo     CI.hi
#> 1  Prop 0.6450265 0.1328846 0.3845775 0.9054755
#> 2  Tmin 2.9341378 0.7320888 1.4992701 4.3690055
#> 3  Gmin 5.3163693 1.4801290 2.4153697 8.2173689

# passing only arguments needed for prop point estimate
confounded_meta( q=log(0.90), muB=log(1.5),
                 yr=yr, t2=t2, CI.level=0.95 )

#> Cannot compute inference without vyr and vt2. Returning only point estimates.

#> Cannot compute Tmin or Gmin without r. Returning only prop.

#>   Value       Est SE CI.lo CI.hi
#> 1  Prop 0.6427633 NA    NA    NA
#> 2  Tmin        NA NA    NA    NA
#> 3  Gmin        NA NA    NA    NA

# passing only arguments needed for Tmin, Gmin point estimates
confounded_meta( q=log(0.90), r=0.20,
                 yr=yr, t2=t2, CI.level=0.95 )

#> Cannot compute inference without vyr and vt2. Returning only point estimates.

#>   Value      Est SE CI.lo CI.hi
#> 1  Prop       NA NA    NA    NA
#> 2  Tmin 2.934138 NA    NA    NA
#> 3  Gmin 5.316369 NA    NA    NA