Unmeasured confounding

A type of bias. Declares that unmeasured confounding will be a component of interest in the multi-bias sensitivity analysis. Generally used within other functions, its output is returned invisibly.

confounding(..., verbose = FALSE)

Arguments

...
Other arguments. Not currently used for this function.

verbose
Logical. If TRUE, returns warnings and messages immediately. Defaults to FALSE because it is generally used within the multi_bias() function, which will print the same messages/warnings.

Value

Invisibly returns a list with components n (2, the degree of the polynomial in the numerator), d (1, the degree of the polynomial in the denominator), mess (any messages/warnings that should be printed for the user), and bias ("confounding").

Examples

# returns invisibly without print()
print(confounding())

#> [1] "confounding"

# Calculate an E-value for unmeasured confounding only
multi_evalue(est = RR(4), biases = confounding())

#> This multi-bias e-value refers simultaneously to parameters RR_AUc,
#> RR_UcY . (See documentation for details.)

#>                        point lower upper
#> RR                  4.000000    NA    NA
#> Multi-bias E-values 7.464102    NA    NA