Compute E-value for a risk ratio or rate ratio and its confidence interval limits

Returns a data frame containing point estimates, the lower confidence limit, and the upper confidence limit for the risk ratio (as provided by the user) as well as E-values for the point estimate and the confidence interval limit closer to the null.

evalues.RR(est, lo = NA, hi = NA, true = 1, ...)

Arguments

est
The point estimate

lo
The lower limit of the confidence interval

hi
The upper limit of the confidence interval

true
The true RR to which to shift the observed point estimate. Typically set to 1 to consider a null true effect.

...
Arguments passed to other methods.

Examples

# compute E-value for leukemia example in VanderWeele and Ding (2017)
evalues.RR(0.80, 0.71, 0.91)

#>             point lower    upper
#> RR       0.800000  0.71 0.910000
#> E-values 1.809017    NA 1.428571

# you can also pass just the point estimate
evalues.RR(0.80)

#>             point lower upper
#> RR       0.800000    NA    NA
#> E-values 1.809017    NA    NA

# demonstrate symmetry of E-value
# this apparently causative association has same E-value as the above
evalues.RR(1 / 0.80)

#>             point lower upper
#> RR       1.250000    NA    NA
#> E-values 1.809017    NA    NA