Calculate a bound for the bias

Function used to calculate the maximum factor by which a risk ratio is biased, given possible values for each of the parameters that describe the bias factors for each type of bias.

multi_bound(
  biases,
  RRAUc = NULL,
  RRUcY = NULL,
  RRUsYA1 = NULL,
  RRSUsA1 = NULL,
  RRUsYA0 = NULL,
  RRSUsA0 = NULL,
  RRAUscS = NULL,
  RRUscYS = NULL,
  RRAYy = NULL,
  ORYAa = NULL,
  RRYAa = NULL,
  RRAYyS = NULL,
  ORYAaS = NULL,
  RRYAaS = NULL,
  RRAUsS = NULL,
  RRUsYS = NULL
)

Arguments

biases
A set of biases (or single bias) to include in the calculation of the bound. A single object constructed with the multi_bias() function, it may include any or all of confounding(), selection(), and misclassification(), and any of the options described in the documentation for those functions.

RRAUc
Named parameter values with which to calculate a bound. Names must correspond to the parameters defining the biases provided by biases. Help with names can be found by running print(multi_bias(...)) for the biases of interest. Unnecessary parameters are ignored with a warning.

RRUcY
See RRAUc

RRUsYA1
See RRAUc

RRSUsA1
See RRAUc

RRUsYA0
See RRAUc

RRSUsA0
See RRAUc

RRAUscS
See RRAUc

RRUscYS
See RRAUc

RRAYy
See RRAUc

ORYAa
See RRAUc

RRYAa
See RRAUc

RRAYyS
See RRAUc

ORYAaS
See RRAUc

RRYAaS
See RRAUc

RRAUsS
See RRAUc

RRUsYS
See RRAUc

Value

Returns the value of the bound formed as a function of the provided parameters.

Details

The names of the parameters in the bound can be found for a given set of biases with print(biases). Running summary(biases) shows the equivalent notation used in the output of the multi_evalue() function.

Examples

multi_bound(multi_bias(confounding()),
            RRAUc = 2.2, RRUcY = 1.7)

#> [1] 1.289655

biases <- multi_bias(confounding(), selection("S = U"),
                     misclassification("exposure",
                     rare_outcome = TRUE, rare_exposure = FALSE))

#> The default option, general selection bias, is being used as well.

print(biases)

#> The following arguments can be copied and pasted into the multi_bound()
#> function: RRAUc = , RRUcY = , RRSUsA1 = , RRSUsA0 = , ORYAaS =

multi_bound(biases,
            RRAUc = 3, RRUcY = 2, RRSUsA1 = 2.3,
            RRSUsA0 = 1.7, ORYAaS = 5.2)

#> [1] 30.498