Calculate a multiple-bias E-value

Calculate an E-value for a specified set of biases.

multi_evalue(biases, est, ...)

multi_evalues.HR(
  biases,
  est,
  lo = NA,
  hi = NA,
  rare = NULL,
  true = 1,
  verbose = TRUE,
  ...
)

multi_evalues.OR(
  biases,
  est,
  lo = NA,
  hi = NA,
  rare = NULL,
  true = 1,
  verbose = TRUE,
  ...
)

multi_evalues.RR(biases, est, lo = NA, hi = NA, true = 1, verbose = TRUE, ...)

Arguments

biases
An object created by multi_bias() (or a single bias) to include in the calculation of the E-value. May include any or all of confounding(), selection(), and misclassification(), and any of the options described in the documentation for those functions.

est
The effect estimate that was observed but which is suspected to be biased. This may be of class "estimate" (constructed with RR(), OR(), or HR(), or more information can be provided using the other arguments.

...
Arguments passed to other methods.

lo
Optional. Lower bound of the confidence interval. If not an object of class "estimate", assumed to be on the same scale as est.

hi
Optional. Upper bound of the confidence interval. If not an object of class "estimate", assumed to be on the same scale as est.

rare
Logical indicating whether outcome is sufficiently rare for risk ratio approximation to hold.

true
A number to which to shift the observed estimate to. Defaults to

If not an object of class "estimate", assumed to be on the same scale as est.

verbose
Logical indicating whether or not to print information about which parameters the multi-bias E-value refers to. Defaults to TRUE.

Value

Returns a multiple bias E-value, of class "multi_evalue", describing the value that each of a number of parameters would have to have for the observed effect measure to be completely explained by bias.

Examples

# Calculate an E-value for unmeasured confounding
multi_evalue(est = RR(4), biases = confounding())

#> This multi-bias e-value refers simultaneously to parameters RR_AUc,
#> RR_UcY . (See documentation for details.)

#>                        point lower upper
#> RR                  4.000000    NA    NA
#> Multi-bias E-values 7.464102    NA    NA

# Equivalent to
evalues.RR(4)

#>             point lower upper
#> RR       4.000000    NA    NA
#> E-values 7.464102    NA    NA

# Calculate a multi-bias E-value for selection bias
# and misclassification
multi_evalue(est = RR(2.5),
         biases = multi_bias(selection("selected"),
                   misclassification("outcome")))

#> This multi-bias e-value refers simultaneously to parameters RR_AUs|S,
#> RR_UsY|S, RR_AY*|y,S . (See documentation for details.)

#>                        point lower upper
#> RR                  2.500000    NA    NA
#> Multi-bias E-values 1.923636    NA    NA

# Calculate a multi-bias E-value for all three
# available types of bias
biases <- multi_bias(confounding(),
                     selection("general", "S = U"),
                     misclassification("exposure",
                            rare_outcome = TRUE))
multi_evalue(est = RR(2.5), biases = biases)

#> This multi-bias e-value refers simultaneously to parameters RR_AUc,
#> RR_UcY, RR_SUs|A=1, RR_SUs|A=0, RR_YA*|a,S . (See documentation for
#> details.)

#>                        point lower upper
#> RR                  2.500000    NA    NA
#> Multi-bias E-values 1.245073    NA    NA

# Calculate a multi-bias E-value for a non-rare OR
# using the square root approximation
multi_evalue(est = OR(2.5, rare = FALSE), biases = biases)

#> This multi-bias e-value refers simultaneously to parameters RR_AUc,
#> RR_UcY, RR_SUs|A=1, RR_SUs|A=0, RR_YA*|a,S . (See documentation for
#> details.)

#>                        point lower upper
#> RR                  1.581139    NA    NA
#> Multi-bias E-values 1.118186    NA    NA

# Calculate a non-null multi-bias E-value
multi_evalue(est = RR(2.5), biases = biases, true = 2)

#> You are calculating a "non-null" multi-bias E-value, i.e., a multi-bias
#> E-value for the minimum amount of bias needed to move the estimate and
#> confidence interval to your specified true value of 2 rather than to
#> the null value.

#> This multi-bias e-value refers simultaneously to parameters RR_AUc,
#> RR_UcY, RR_SUs|A=1, RR_SUs|A=0, RR_YA*|a,S . (See documentation for
#> details.)

#>                        point lower upper
#> RR                  2.500000    NA    NA
#> Multi-bias E-values 1.056594    NA    NA