Estimate proportion of population effect sizes above or below a threshold

Estimates the proportion of true effect sizes in a meta-analysis above or below a specified threshold of scientific importance. Effect sizes may be of any type (they need not be relative risks). This is a wrapper for confounded_meta; it is the special case in which there is no unmeasured confounding.

stronger_than(q, yr, vyr = NA, t2, vt2 = NA, CI.level = 0.95, tail)

Arguments

q
True effect size that is the threshold for "scientific importance"

yr
Pooled point estimate from meta-analysis

vyr
Estimated variance of pooled point estimate from meta-analysis

t2
Estimated heterogeneity (tau^2) from meta-analysis

vt2
Estimated variance of tau^2 from meta-analysis

CI.level
Confidence level as a proportion

tail
above for the proportion of effects above q; below for the proportion of effects below q.